requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
Deprecated==1.2.13
Django==4.0.6
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-mdeditor==0.1.20
django-redis==5.2.0
importlib-metadata==5.0.0
Markdown==3.4.1
packaging==21.3
PyMySQL==1.0.2
pyparsing==3.0.9
redis==4.3.4
sqlparse==0.4.3
tzdata==2022.6
wrapt==1.14.1
zipp==3.10.0
gunicorn==20.1.0
gevent==21.12.0